Updated error.py with relevant files
