+++ error.py
@@ -35,5 +35,22 @@ class WebsiteNotStampifiableError(StampifierError):
 class WebsiteConnectionError(StampifierError):
     """Raise when connection to website is not possible"""
 
-    def __init__(self):
-        super().__init__('Cannot connect to URL!')
+    def __init__(self, url):
+        super().__init__('Cannot connect to URL! {}'.format(url))
+
+
+class BadRequestError(Exception):
+    '''Exception raised when the API call was not completed successfully'''
+
+    def __init__(self, status_code):
+        super(BadRequestError, self).__init__()
+        self.message = "The API call was unsuccessful with status code: " + \
+            str(status_code)
+
+
+class IncorrectInputError(Exception):
+    ''' Exception raised when the input format is wrong'''
+    def __init__(self, message):
+        super(IncorrectInputError, self).__init__()
+        self.message = message
+
