--- summarization/incorrect_input.py
@@ -1,8 +1,0 @@
-''' Class to define the IncorrectInputException'''
-
-
-class IncorrectInputError(Exception):
-    ''' Exception raised when the input format is wrong'''
-    def __init__(self, message):
-        super(IncorrectInputError, self).__init__()
-        self.message = message
