+++ summarization/web_entity_detection.py
@@ -18,8 +18,8 @@ import os
 import requests
 from nltk.tokenize import word_tokenize
 
-from summarization.bad_request_error import BadRequestError
 from utils.url_utils import convert_scheme_to_http
+from error import BadRequestError
 
 
 class ImageDescriptionRetriever:
