+++ tests/summarizer/test_text_entity_retriever.py
@@ -3,7 +3,7 @@ from unittest.mock import Mock, patch
 
 import pytest
 
-from summarization.bad_request_error import BadRequestError
+from error import BadRequestError
 from summarization.text_entity_detection import TextEntityRetriever
